requirements.txt
llama-cpp-python
langchain
sentence-transformers
faiss-cpu
streamlit
langchain-community
pypdf
psutil
